Move subcommands under src. Otherwise, pub-global-activate believes they should have a main
