+++ bin/src/coverage_log_server.dart
