+++ bin/src/debug_info.dart
