+++ bin/src/deferred_library_check.dart
