+++ bin/src/deferred_library_size.dart
