+++ bin/src/diff.dart
