+++ bin/src/inject_text.dart
