+++ bin/src/library_size_split.dart
