+++ bin/src/to_binary.dart
